+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @ballerine/headless-example
 
+## 0.3.132
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.110
+  - @ballerine/workflow-browser-sdk@0.6.133
+
+## 0.3.131
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.109
+  - @ballerine/workflow-browser-sdk@0.6.132
+
+## 0.3.130
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.108
+  - @ballerine/workflow-browser-sdk@0.6.131
+
+## 0.3.129
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.107
+  - @ballerine/workflow-browser-sdk@0.6.130
+
 ## 0.3.128
 
 ### Patch Changes