+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @ballerine/workflow-core
 
+## 0.6.133
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.110
+
+## 0.6.132
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.109
+
+## 0.6.131
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.108
+
+## 0.6.130
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.107
+
 ## 0.6.129
 
 ### Patch Changes