+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,34 @@
 # web-ui-sdk
 
+## 1.5.111
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.110
+
+## 1.5.110
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.109
+
+## 1.5.109
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.108
+
+## 1.5.108
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.107
+
 ## 1.5.107
 
 ### Patch Changes