+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.133
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.110
+  - @ballerine/workflow-core@0.6.133
+
+## 0.6.132
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.109
+  - @ballerine/workflow-core@0.6.132
+
+## 0.6.131
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.108
+  - @ballerine/workflow-core@0.6.131
+
+## 0.6.130
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.107
+  - @ballerine/workflow-core@0.6.130
+
 ## 0.6.129
 
 ### Patch Changes