+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.133
+
+### Patch Changes
+
+- Bump
+  - @ballerine/workflow-core@0.6.133
+
+## 0.6.132
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.132
+
+## 0.6.131
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.131
+
+## 0.6.130
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.130
+
 ## 0.6.129
 
 ### Patch Changes
@@ -13,6 +40,7 @@
 - bump
 - Updated dependencies
   - @ballerine/workflow-core@0.6.128
+- @ballerine/workflow-core@0.6.128
 
 ## 0.6.127